Avoid TSan data race with CompressionState
Most of CompressionState is only mutated once per connection, as
part of setting up the websocket.

But needsCompression is mutated per-read-message.

Handling reads is serialized on one queue, and handling writes is serialized on
another queueu, which means that a read can be handled concurrently with
handling a write.

In theory this concurrent read/write should be harmless because the mutated
field is not accessed by the other queue, but for best practice, we should
write our code in a way that keeps TSan happy.